@@ -6,20 +6,20 @@ description: Demonstrates how to use the flutter_nearby_connections plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  fluttertoast: ^7.0.4
+  flutter_styled_toast: ^2.0.0
 
   flutter_nearby_connections:
     path: ../
 
-  cupertino_icons: ^0.1.3
-  device_info: ^0.4.0+4
-  flutter_udid: ^1.0.1
+  cupertino_icons: ^1.0.2
+  device_info: ^2.0.0
+  flutter_udid: ^2.0.0
 
 dev_dependencies:
   flutter_test:
