@@ -1,19 +1,18 @@
 name: flutter_nearby_connections
 description: Flutter plugin supports peer-to-peer connectivity and discovers nearby devices for Android and IOS
-version: 1.0.10
+version: 1.1.1
 author: vn.apnic@gmail.com
 homepage: https://github.com/VNAPNIC/flutter_nearby_connections
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  pedantic: ^1.9.0
-  fluttertoast: ^7.1.1
   rxdart: 0.24.1
+
 
 dev_dependencies:
   flutter_test:
